Remove CSS-in-JS index page generation. Also rename and relocate files.
