--- src/setup/steps/add-content/generate-index/with-css-modules/generate-index-with-css-modules.ts
@@ -1,63 +1,0 @@
-import { ValidCNSInputs } from "../../../../../create-next-stack-types"
-import { getProjectNameOfPath } from "../../../../../helpers/get-project-name-of-path"
-import {
-  confettiImports,
-  confettiScriptComponent,
-  onConfettiLoadFunction,
-} from "../confetti-script"
-
-export const generateIndexWithCssModules = (
-  inputs: ValidCNSInputs
-) => /* tsx */ `
-import { NextPage } from "next";
-${confettiImports}
-import Page from "../components/Page";
-import styles from "../styles/index.module.css";
-
-const Index: NextPage = () => {
-  ${onConfettiLoadFunction}
-
-  return (
-    <Page
-      title="${getProjectNameOfPath(inputs.args.appName)}"
-      description="Generated by Create Next Stack."
-    >
-      ${confettiScriptComponent}
-      <main>
-        <section className={styles.pageSection}>
-          <div className={styles.contentContainer}>
-            <h1 className={styles.title}>Your project is a go! 🎉</h1>
-            <p className={styles.subtitle}>
-              Get started by editing{" "}
-              <code className={styles.inlineCode}>pages/index.tsx</code>
-            </p>
-          </div>
-        </section>{" "}
-        <section className={\`\${styles.pageSection} \${styles.grayBackground}\`}>
-          <div className={styles.contentContainer}>
-            <h2 className={styles.h2}>Learning resources</h2>
-            <p className={styles.paragraph}>
-              If you are using a technology for the first time, you can find
-              related links in the generated{" "}
-              <code className={styles.inlineCode}>README.md</code> file that
-              might prove helpful.
-            </p>
-          </div>
-        </section>
-      </main>
-      <footer className={styles.pageSection}>
-        <div className={styles.contentContainer}>
-          <p className={styles.paragraph}>
-            Generated by{" "}
-            <a href="https://github.com/akd-io/create-next-stack">
-              Create Next Stack
-            </a>
-          </p>
-        </div>
-      </footer>
-    </Page>
-  );
-};
-
-export default Index;
-`
