+++ src/setup/steps/add-content/index-page/confetti-script.ts
